backend/requirements.txt
asgiref==3.8.1
boto3==1.35.89
botocore==1.35.89
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
Django==5.1.4
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==24.3
django-otp==1.5.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
jmespath==1.0.1
localstack-client==2.7
Markdown==3.7
MarkupSafe==3.0.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
rest-framework-simplejwt==0.0.2
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
